BAG-55: structure stack for the recipes catalogue, re-organize folders
