+++ App.js
@@ -1,9 +1,9 @@
 import 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import PlanScreen from './screens/PlanScreen';
-import ListScreen from './screens/ListScreen';
+import ListScreen from './screens/planning screens/ListScreen';
 import HomeScreen from './screens/HomeScreen';
+import RecipeScreen from './screens/recipe catalogue screens/RecipeScreen';
 
 const Stack = createStackNavigator();
 
@@ -12,7 +12,12 @@ export default function App() {
     <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen name="Home" component={HomeScreen} options={{ title: 'Plan', headerShown: false }}/>
+
+        {/* Planning Screen */}
         <Stack.Screen name="List" component={ListScreen} />
+
+        {/* Recipes Catalogue */}
+        <Stack.Screen name="Recipe" component={RecipeScreen} />
       </Stack.Navigator>
     </NavigationContainer>
   );
