+++ screens/HomeScreen.js
@@ -1,9 +1,9 @@
 import { View, Text } from 'react-native'
 import React from 'react'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import PlanScreen from './PlanScreen';
-import RecipesScreen from './RecipesScreen';
-import StoreScreen from './StoreScreen';
+import PlanScreen from './planning screens/PlanScreen';
+import RecipesScreen from './recipe catalogue screens/RecipesScreen';
+import StoreScreen from './store managing screens/StoreScreen';
 
 const Tab = createBottomTabNavigator();
 
@@ -11,7 +11,7 @@ const HomeScreen = () => {
   return (
     <Tab.Navigator>
         <Tab.Screen name="Plan" component={PlanScreen} />
-        <Tab.Screen name="Recipe" component={RecipesScreen} />
+        <Tab.Screen name="Recipes" component={RecipesScreen} />
         <Tab.Screen name="Store" component={StoreScreen} />
     </Tab.Navigator>
   )
