+++ screens/recipe catalogue screens/RecipesScreen.js
@@ -1,0 +1,16 @@
+import { View, Text, Pressable } from 'react-native'
+import React from 'react'
+import styles from '../../styles/Styles'
+
+const RecipesScreen = ({ navigation }) => {
+  return (
+    <View style={styles.container}>
+      <Text>RecipesScreen</Text>
+      <Pressable onPress={() => navigation.navigate("Recipe")}>
+        <Text>Go to recipe</Text>
+      </Pressable>
+    </View>
+  )
+}
+
+export default RecipesScreen
