+++ screens/store managing screens/StoreScreen.js
@@ -1,13 +1,13 @@
 import { View, Text } from 'react-native'
 import React from 'react'
-import styles from '../styles/Styles'
+import styles from '../../styles/Styles'
 
-const ListScreen = () => {
+const StoreScreen = () => {
   return (
     <View style={styles.container}>
-      <Text>ListScreen</Text>
+      <Text>StoreScreen</Text>
     </View>
   )
 }
 
-export default ListScreen
+export default StoreScreen
